Various updates
Added license header, updated documentation. Also, added a new feature
that prevents multiple hits when holding down a laser. Previously, it
would play the sound over and over, which is OK for a wind instrument,
but horrible for string and piano instruments.

This commit adds a variable and related code to keep track of the last
hit laser, which causes it to play the sound once, even if the laser is
held down.

Also added are some #ifndefs that disable the new code when #define WIND
is uncommented. This reverts this program back to the old behavior.