Support for multiple strings in multiple hit detection
Previously, the instrument would revert to the old behavior of repeatedly
playing the same notes when holding a string down, which is wrong. This
was a limitation of the way that we tracked the last held string, which
only supports tracking one string at a time. However, when you held two
strings, what would happen is the last string tracker would switch back
and forth, rendering it useless.

This commit replaces the old system with a new one based on bitwise
operations and flags to keep track. The last string tracker is now just
a tracker in general, as it is an unsigned int. Each laser has it's own
flag assigned, with laser1 as 0x01, laser2 and 0x02, laser3 as 0x04, and
so on. Then, we can set the flags with a bitwise OR operator, and read
the flags with the AND operator. This way, we really are tracked each laser
individually, but instead of using 8 variables, we use one.

This code is experimental, and needs LOTS of testing.