Use array for instruments
Use an array to store instruments, allowing us to cycle between instruments
with the press of a button. The nice thing is you can add more instruments
relatively easily, as long as timidity supports it.

If timidity has no output, you may need to install an extended sound font
for timidty, such as Fluid General MIDI SoundFont

This commit also fixes some glitches with button presses, although this code
is likely to not work, because it has not been tested on real hardware.